canaries/unsloth/requirements.txt
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
torch>=2.4.0
transformers>=4.44.0
datasets>=2.20.0
trl>=0.9.0
peft>=0.12.0
bitsandbytes>=0.43.0
pyyaml>=6.0
